# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arc LMS API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Arc LMS Support",
      "email": "support@arc-lms.com",
      "url": "https://arc-lms.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://arc-lms.onrender.com/api/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/courses/{id}/quizzes": {
      "get": {
        "tags": ["Assessments"],
        "summary": "List quizzes for a course",
        "operationId": "listQuizzes",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Quizzes retrieved", "content": {"application/json": {"schema": {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/Quiz"}}}}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      },
      "post": {
        "tags": ["Assessments"],
        "summary": "Create quiz",
        "operationId": "createQuiz",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateQuizRequest"}}}},
        "responses": {
          "201": {"description": "Quiz created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Quiz"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/courses/{id}/assignments": {
      "get": {
        "tags": ["Assessments"],
        "summary": "List assignments for a course",
        "operationId": "listAssignments",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Assignments retrieved", "content": {"application/json": {"schema": {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/Assignment"}}}}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      },
      "post": {
        "tags": ["Assessments"],
        "summary": "Create assignment",
        "operationId": "createAssignment",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateAssignmentRequest"}}}},
        "responses": {
          "201": {"description": "Assignment created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Assignment"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /public/auth/login"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "description": "Unique key to ensure idempotent requests. Prevents duplicate operations.",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required or token invalid/expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required",
                "details": {}
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "FORBIDDEN",
                "message": "Insufficient permissions to perform this action",
                "details": {}
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Quiz": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string", "format": "uuid"},
          "course_id": {"type": "string", "format": "uuid"},
          "title": {"type": "string"},
          "description": {"type": "string"},
          "instructions": {"type": "string"},
          "questions": {"type": "array", "items": {"$ref": "#/components/schemas/Question"}},
          "time_limit_minutes": {"type": "integer"},
          "max_attempts": {"type": "integer"},
          "passing_score": {"type": "number"},
          "total_points": {"type": "number"},
          "shuffle_questions": {"type": "boolean"},
          "shuffle_options": {"type": "boolean"},
          "show_results": {"type": "boolean"},
          "availability_start": {"type": "string", "format": "date-time"},
          "availability_end": {"type": "string", "format": "date-time"},
          "status": {"type": "string", "enum": ["DRAFT", "PUBLISHED", "CLOSED", "ARCHIVED"]},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "CreateQuizRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {"type": "string", "minLength": 1, "maxLength": 200},
          "description": {"type": "string", "maxLength": 1000},
          "instructions": {"type": "string", "maxLength": 2000},
          "questions": {"type": "array", "items": {"$ref": "#/components/schemas/QuestionInput"}},
          "time_limit_minutes": {"type": "integer", "minimum": 1},
          "max_attempts": {"type": "integer", "minimum": 1},
          "passing_score": {"type": "number", "minimum": 0, "maximum": 100},
          "shuffle_questions": {"type": "boolean"},
          "shuffle_options": {"type": "boolean"},
          "show_results": {"type": "boolean"},
          "availability_start": {"type": "string", "format": "date-time"},
          "availability_end": {"type": "string", "format": "date-time"}
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "type": {"type": "string", "enum": ["MULTIPLE_CHOICE", "TRUE_FALSE", "SHORT_ANSWER", "ESSAY", "FILL_IN_BLANK"]},
          "text": {"type": "string"},
          "options": {"type": "array", "items": {"$ref": "#/components/schemas/QuestionOption"}},
          "correct_answer": {"type": "string"},
          "points": {"type": "number"},
          "explanation": {"type": "string"},
          "order": {"type": "integer"}
        }
      },
      "QuestionInput": {
        "type": "object",
        "required": ["type", "text", "points"],
        "properties": {
          "type": {"type": "string", "enum": ["MULTIPLE_CHOICE", "TRUE_FALSE", "SHORT_ANSWER", "ESSAY", "FILL_IN_BLANK"]},
          "text": {"type": "string", "minLength": 1},
          "options": {"type": "array", "items": {"$ref": "#/components/schemas/QuestionOptionInput"}},
          "correct_answer": {"type": "string"},
          "points": {"type": "number", "minimum": 0},
          "explanation": {"type": "string"}
        }
      },
      "QuestionOption": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "text": {"type": "string"},
          "is_correct": {"type": "boolean"}
        }
      },
      "QuestionOptionInput": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {"type": "string", "minLength": 1},
          "is_correct": {"type": "boolean"}
        }
      },
      "Assignment": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string", "format": "uuid"},
          "course_id": {"type": "string", "format": "uuid"},
          "title": {"type": "string"},
          "description": {"type": "string"},
          "instructions": {"type": "string"},
          "due_date": {"type": "string", "format": "date-time"},
          "total_points": {"type": "number"},
          "allow_late_submission": {"type": "boolean"},
          "late_penalty_percent": {"type": "number"},
          "attachment_urls": {"type": "array", "items": {"type": "string", "format": "uri"}},
          "status": {"type": "string", "enum": ["DRAFT", "PUBLISHED", "CLOSED", "ARCHIVED"]},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "CreateAssignmentRequest": {
        "type": "object",
        "required": ["title", "due_date", "total_points"],
        "properties": {
          "title": {"type": "string", "minLength": 1, "maxLength": 200},
          "description": {"type": "string", "maxLength": 2000},
          "instructions": {"type": "string", "maxLength": 5000},
          "due_date": {"type": "string", "format": "date-time"},
          "total_points": {"type": "number", "minimum": 0},
          "allow_late_submission": {"type": "boolean"},
          "late_penalty_percent": {"type": "number", "minimum": 0, "maximum": 100},
          "attachment_urls": {"type": "array", "items": {"type": "string", "format": "uri"}}
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error context",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
    }
  }
}